+++ src/MP_Blynk_ESP8266_Serial1.cpp
@@ -5,7 +5,7 @@
 #include "MP_Blynk_ESP8266_Serial1.h"
 #include <BlynkSimpleShieldEsp8266.h>
 
-int MP_Blynk_ESP8266_Serial1::value[8];
+double MP_Blynk_ESP8266_Serial1::value[8];
 uint8_t MP_Blynk_ESP8266_Serial1::valueChanged;
 
 MP_Blynk_ESP8266_Serial1::MP_Blynk_ESP8266_Serial1(char* auth, char* ssid, char* pass)
@@ -51,7 +51,7 @@ int MP_Blynk_ESP8266_Serial1::readVirtualPin(uint8_t pin)
     return MP_Blynk_ESP8266_Serial1::value[pin];
 }
 
-void MP_Blynk_ESP8266_Serial1::writeVirtualPin(char pin[], int value)
+void MP_Blynk_ESP8266_Serial1::writeVirtualPin(char pin[], double value)
 {
     MP_Blynk_ESP8266_Serial1::value[pin[0] - '0'] = value;
     MP_Blynk_ESP8266_Serial1::valueChanged |= (1 << (pin[0] - '0'));
